+++ bingads/v13/bulk/entities/__init__.py
@@ -27,3 +27,4 @@ from .target_criterions import *
 from .labels import *
 from .bulk_offline_conversion import *
 from .bulk_experiment import *
+from .feeds import *